feat: add header getFormattedFieldValue method
